requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
astunparse==1.6.3
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-PyMongo==3.0.1
flatbuffers==25.2.10
fsspec==2025.5.1
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.53.0
h5py==3.14.0
httplib2==0.22.0
huggingface-hub==0.33.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==2.15.0
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==3.0.2
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==11.2.1
proto-plus==1.26.1
protobuf==4.23.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.2
pyparsing==3.2.3
pypdfium2==4.30.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.4.2
scipy==1.13.1
six==1.17.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.15.2
tqdm==4.67.1
transformers==4.39.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.23.0
